10035 move and simplify versioncontrolheader
