+++ frontend/packages/shared/src/api-paths/cache-key.ts
@@ -6,4 +6,6 @@ export enum CacheKey {
   AppDeployments = 'APP_DEPLOYMENTS',
   BranchStatus = 'BRANCH_STATUS',
   Environemnts = 'ENVIRONMENTS',
+  RepoMetaData = 'REPO_METADATA',
+  RepoPullData = 'REPO_PULL_DATA',
 }
